+++ src/main/java/com/ibrasoft/lensbridge/repository/auth/RefreshTokenRepository.java
@@ -1,7 +1,8 @@
-package com.ibrasoft.lensbridge.repository.mongo;
+package com.ibrasoft.lensbridge.repository.auth;
 
 import com.ibrasoft.lensbridge.model.auth.RefreshToken;
-import org.springframework.data.mongodb.repository.MongoRepository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
@@ -10,7 +11,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface RefreshTokenRepository extends MongoRepository<RefreshToken, UUID> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, UUID> {
     
     Optional<RefreshToken> findByToken(String token);
