+++ src/main/java/com/ibrasoft/lensbridge/repository/auth/UserRepository.java
@@ -1,15 +1,15 @@
-package com.ibrasoft.lensbridge.repository.mongo;
+package com.ibrasoft.lensbridge.repository.auth;
 
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.ibrasoft.lensbridge.model.auth.User;
 
 @Repository
-public interface UserRepository extends MongoRepository<User, UUID> {
+public interface UserRepository extends JpaRepository<User, UUID> {
 
   Optional<User> findByEmail(String email);
