+++ src/main/java/com/ibrasoft/lensbridge/repository/auth/VerificationTokenRepository.java
@@ -1,0 +1,14 @@
+package com.ibrasoft.lensbridge.repository.auth;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.ibrasoft.lensbridge.model.auth.User;
+import com.ibrasoft.lensbridge.model.auth.VerificationToken;
+
+@Repository
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, UUID> {
+    User findUserByToken(String tokenHash);
+}
